+++ src/datasentry/__init__.py
@@ -2,6 +2,6 @@
 
 from datasentry.client import DataSentry
 
-__version__ = "0.27.1"
+__version__ = "0.28.0"
 
 __all__ = ["DataSentry", "__version__"]